requirements.txt
streamlit==1.55.0
python-dotenv==1.2.2
langchain-community==0.4.1
langchain-openai==1.1.11
langchain-classic==1.0.3
langchain-text-splitters==1.1.1
faiss-cpu==1.13.2
sentence-transformers==5.3.0
tiktoken==0.12.0
pypdf==6.8.0
docx2txt==0.9
unstructured==0.18.32
openpyxl==3.1.5